data/requirements.txt
kagglehub
llama-index
weaviate-client
llama-index-vector-stores-weaviate
llama-index-embeddings-text-embeddings-inference
pandas